# backend/contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/files/upload-url": {
      "post": {
        "operationId": "FilesController_uploadUrl",
        "summary": "Get a signed upload URL + fileKey",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/files/confirm": {
      "post": {
        "operationId": "FilesController_confirm",
        "summary": "Persist a file row after upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/files/{id}/download-url": {
      "get": {
        "operationId": "FilesController_downloadUrl",
        "summary": "Get a signed download URL (ownership checked)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/files/blob": {
      "put": {
        "operationId": "FilesController_putBlob",
        "summary": "[local] Upload bytes to a signed URL",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ]
      },
      "get": {
        "operationId": "FilesController_getBlob",
        "summary": "[local] Download bytes from a signed URL",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ]
      }
    },
...
    "/api/v1/conversations": {
      "get": {
        "operationId": "ChatController_list",
        "summary": "Thread list (last message + unread count)",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ChatController_create",
        "summary": "Get-or-create a 1:1 conversation",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/conversations/realtime-token": {
      "get": {
        "operationId": "ChatController_realtimeToken",
        "summary": "Credentials for the RealtimePort",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/conversations/{id}/messages": {
      "get": {
        "operationId": "ChatController_messages",
        "summary": "Paginated message history (cursor)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ChatController_send",
        "summary": "Send a message",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/conversations/{id}/read": {
      "post": {
        "operationId": "ChatController_read",
        "summary": "Mark conversation read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/conversations/stream": {
      "get": {
        "operationId": "ChatController_stream",
        "summary": "[local] SSE event stream",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "Refiqo API",
    "description": "Refiqo REST API — referral networking platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}